+++ doc/CHANGES.rst
@@ -13,6 +13,9 @@ Bugs Fixed
 
 - Fixed two unit tests that failed on fast Windows machines.
 
+- Fixed OverflowError in Products.ZCatalog.Lazy on 64bit python
+  on Windows.
+
 - Fixed ``testZODBCompat`` tests in ZopeTestCase to match modern ZODB
   semantics.